+++ libsrc/ffdec_lib/src/com/jpexs/decompiler/flash/exporters/settings/ScriptExportSettings.java
@@ -12,7 +12,8 @@
  * Lesser General Public License for more details.
  * 
  * You should have received a copy of the GNU Lesser General Public
- * License along with this library. */
+ * License along with this library.
+ */
 package com.jpexs.decompiler.flash.exporters.settings;
 
 import com.jpexs.decompiler.flash.exporters.modes.ScriptExportMode;
@@ -32,9 +33,12 @@ public class ScriptExportSettings {
 
     public FileTextWriter singleFileWriter;
 
-    public ScriptExportSettings(ScriptExportMode mode, boolean singleFile) {
+    public boolean ignoreFrameScripts;
+
+    public ScriptExportSettings(ScriptExportMode mode, boolean singleFile, boolean ignoreFrameScripts) {
         this.mode = mode;
         this.singleFile = singleFile;
+        this.ignoreFrameScripts = ignoreFrameScripts;
     }
 
     public String getFileExtension() {